@@ -5,7 +5,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.19.0 <3.0.0"
+  sdk: '>=3.0.1 <4.0.0'
 
 dependencies:
   flutter:
@@ -21,6 +21,7 @@ dependencies:
   uuid: ^3.0.7
   http: ^0.13.5
   get_it: ^7.2.0
+  date_time_picker: ^2.1.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
