@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:intl/intl.dart';
+import 'package:flutter/services.dart';
 import 'package:personal_expanses_app/currency_api_bloc.dart';
 import 'package:personal_expanses_app/map_view.dart';
 import 'package:get_it/get_it.dart';
@@ -18,6 +18,15 @@ class _NewTransactionState extends State<NewTransaction> {
   final _amountController = TextEditingController();
   DateTime _selectedDate = DateTime.now();
   LocalizationObject? transactionAdress;
+  List<_DateObject> daysOfWeek = List.generate(
+    Days.values.length,
+    (index) => _DateObject(
+      Days.values[index],
+      DateTime.now().subtract(
+        Duration(days: index),
+      ),
+    ),
+  );
 
   void _submitData() {
     if (_amountController.text.isEmpty) {
@@ -29,7 +38,6 @@ class _NewTransactionState extends State<NewTransaction> {
     if (enteredTitle.isEmpty || enteredAmount <= 0) {
       return;
     }
-
     widget.addExpanse(
       enteredTitle,
       enteredAmount,
@@ -40,26 +48,11 @@ class _NewTransactionState extends State<NewTransaction> {
     Navigator.of(context).pop();
   }
 
-  void _presentDatePicker() {
-    showDatePicker(
-      context: context,
-      initialDate: DateTime.now(),
-      firstDate: DateTime(2019),
-      lastDate: DateTime.now(),
-    ).then((pickedDate) {
-      if (pickedDate == null) {
-        return;
-      }
-      setState(() {
-        _selectedDate = pickedDate;
-      });
-    });
-  }
+  _DateObject groupValue = _DateObject(Days.values[DateTime.now().weekday - 1], DateTime.now());
 
   @override
   Widget build(BuildContext context) {
     return Card(
-      key: Key('TestClose'),
       elevation: 8,
       child: Padding(
         padding: EdgeInsets.all(16),
@@ -79,6 +72,7 @@ class _NewTransactionState extends State<NewTransaction> {
                 Expanded(
                   child: TextField(
                     key: Key('AmountInput'),
+                    inputFormatters: [FilteringTextInputFormatter.digitsOnly],
                     decoration: InputDecoration(labelText: 'Amount'),
                     controller: _amountController,
                     keyboardType: TextInputType.number,
@@ -88,47 +82,51 @@ class _NewTransactionState extends State<NewTransaction> {
                 Text(GetIt.I.get<CurrencyApiBloc>().currentCurrencySink.value.name),
               ],
             ),
-            Padding(
-              padding: const EdgeInsets.symmetric(vertical: 8.0),
-              child: Align(
-                alignment: Alignment.centerLeft,
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Text(transactionAdress?.localizationString ?? 'Choose localization'),
-                    IconButton(
-                      onPressed: () async {
-                        final localization = await Navigator.of(context).push<LocalizationObject>(
-                          MaterialPageRoute(builder: (_) => SelectionMapView()),
-                        );
-                        setState(() {
-                          transactionAdress = localization;
-                        });
-                      },
-                      icon: Icon(Icons.map_outlined),
-                    )
-                  ],
+            Visibility(
+              visible: false, // Disabled because of problems with googleApi
+              child: Padding(
+                padding: const EdgeInsets.symmetric(vertical: 8.0),
+                child: Align(
+                  alignment: Alignment.centerLeft,
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      Text(transactionAdress?.localizationString ?? 'Choose localization'),
+                      IconButton(
+                        onPressed: () async {
+                          final localization = await Navigator.of(context).push<LocalizationObject>(
+                            MaterialPageRoute(builder: (_) => SelectionMapView()),
+                          );
+                          setState(() {
+                            transactionAdress = localization;
+                          });
+                        },
+                        icon: Icon(Icons.map_outlined),
+                      )
+                    ],
+                  ),
                 ),
               ),
             ),
-            SizedBox(
-              height: 50,
-              child: Row(
-                children: <Widget>[
-                  Expanded(
-                    child: Text(
-                      'Picked Date: ${DateFormat.yMd().format(_selectedDate)}',
-                    ),
-                  ),
-                  MaterialButton(
-                    textColor: Theme.of(context).primaryColor,
-                    child: Text(
-                      'Choose Date',
-                      style: TextStyle(fontWeight: FontWeight.bold),
-                    ),
-                    onPressed: _presentDatePicker,
-                  ),
-                ],
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: List<Column>.generate(
+                daysOfWeek.length,
+                (index) {
+                  return Column(
+                    children: [
+                      Radio<_DateObject>(
+                        value: daysOfWeek[index],
+                        groupValue: groupValue,
+                        onChanged: (_DateObject? value) {
+                          setState(() => groupValue = value!);
+                          _selectedDate = DateTime.now().subtract(Duration(days: index));
+                        },
+                      ),
+                      Text(Days.values[index].name),
+                    ],
+                  );
+                },
               ),
             ),
             Padding(
@@ -150,3 +148,12 @@ class _NewTransactionState extends State<NewTransaction> {
     );
   }
 }
+
+class _DateObject {
+  _DateObject(this.name, this.dayDate);
+
+  final Days name;
+  final DateTime dayDate;
+}
+
+enum Days { Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday }
